HACK: proc: spoof verifiedbootstate and verity mode
This hack was modified by me a bit from the original author. I added
the option to fully erase a parameter from the cmdline.

This commit also puts the new method into action by using it to spoof
verifiedbootstate to be green (locked and clean), and it also spoofs
veritymode to be enforcing. This prevents the system from finding
anything suspicious, and preventing the user from accessing stuff.