PM / suspend: Clear wakeups before running PM callbacks
PM callbacks can be used as an indicator that the system is suspending, and
as such, a wakeup may be issued outside the notifier in order to cancel an
ongoing suspend. Clearing wakeups after running the PM callbacks can thus
cause wakeups to be missed. Fix it by simply clearing wakeups prior to
running PM callbacks.

Change-Id: I8c438e08bc58afdea8a715d8adb6b3487844bdae
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>